Modified Bruteforce Algorithm for Benes
* Now we try to retrieve the correct status of the selectors
from the measurement and, if there's some sort of uncertainty,
we try to generate the selectors state from the flip measurements.
* Added support function to generate selectors from flip state
* Added support function to get a status array from a probability
array. This function insert a '?' in the array when the bit
status at the corresponding position is not known for sure.
It also computes a sort of accuracy using the maximum of the
probabilities (maybe it needs some refining).